Improve learning objectives for the Terraform and Azure module (#11)
More detailed Terraform learning objectives. Even though we say the main goal of the week is to learn Terraform, the learning objectives don't include any specific goals around Terraform. This PR attempts to rectify that.

This PR is paired with some changes to the Trello board the students have to work through that week.
This is the new version: https://trello.com/b/k4K4ctyt/terraform-and-azure-v2
This is the old version: https://trello.com/b/1w6Va5g0/terraform-week-4

Side-note: eventually it would be better to move these tasks out of Trello entirely and into Slug.

The sequencing of the tasks is still not great. Instead of asking students to create random resources on Azure to allow Terraform state to be stored in them without them understanding why, we should have more material that focuses on understanding the building blocks of terraform and why terraform "state" is even a thing.